✨ Moves interceptors in their own project
