+++ src/Configuration.Persistence.Interceptors/SaveChanges/AuditSaveChangesInterceptor.cs
@@ -1,18 +1,20 @@
-namespace Kritikos.Configuration.Persistence.Interceptors
+namespace Kritikos.Configuration.Persistence.Interceptors.SaveChanges
 {
   using System;
   using System.Diagnostics.CodeAnalysis;
+  using System.Linq;
   using System.Threading;
   using System.Threading.Tasks;
 
   using Kritikos.Configuration.Persistence.Contracts.Behavioral;
-  using Kritikos.Configuration.Persistence.Services;
+  using Kritikos.Configuration.Persistence.Interceptors.Services;
 
   using Microsoft.EntityFrameworkCore;
+  using Microsoft.EntityFrameworkCore.ChangeTracking;
   using Microsoft.EntityFrameworkCore.Diagnostics;
 
   /// <summary>
-  /// Populates audit values for <see cref="IAuditable{T}"/> entities.
+  /// Populates audit values for <see cref="IAuditable{T}"/>, <see cref="ICreateAuditable{T}"/> and <see cref="IUpdateAuditable{T}"/> entities.
   /// </summary>
   /// <typeparam name="T">Type of audit field.</typeparam>
   public class AuditSaveChangesInterceptor<T> : SaveChangesInterceptor
@@ -20,6 +22,10 @@ namespace Kritikos.Configuration.Persistence.Interceptors
   {
     private readonly IAuditorProvider<T> auditorProvider;
 
+    /// <summary>
+    /// Constructs a new instalce of <see cref="AuditSaveChangesInterceptor{T}"/>.
+    /// </summary>
+    /// <param name="auditorProvider">Service that will provide auditors.</param>
     public AuditSaveChangesInterceptor(IAuditorProvider<T> auditorProvider) =>
       this.auditorProvider = auditorProvider;
 
@@ -32,18 +38,7 @@ namespace Kritikos.Configuration.Persistence.Interceptors
       InterceptionResult<int> result)
     {
       var auditor = auditorProvider.GetAuditor() ?? auditorProvider.GetFallbackAuditor();
-      var entries = eventData.Context.ChangeTracker.Entries<IAuditable<T>>()
-                    ?? throw new ArgumentNullException(nameof(eventData));
-
-      foreach (var entry in entries)
-      {
-        if (entry.State == EntityState.Added)
-        {
-          entry.Entity.CreatedBy = auditor;
-        }
-
-        entry.Entity.UpdatedBy = auditor;
-      }
+      StampEntities(eventData.Context.ChangeTracker, auditor);
 
       return base.SavingChanges(eventData, result);
     }
@@ -55,22 +50,28 @@ namespace Kritikos.Configuration.Persistence.Interceptors
       CancellationToken cancellationToken = default)
     {
       var auditor = auditorProvider.GetAuditor() ?? auditorProvider.GetFallbackAuditor();
-      var entries = eventData.Context.ChangeTracker.Entries<IAuditable<T>>()
-                    ?? throw new ArgumentNullException(nameof(eventData));
-
-      foreach (var entry in entries)
-      {
-        if (entry.State == EntityState.Added)
-        {
-          entry.Entity.CreatedBy = auditor;
-        }
-
-        entry.Entity.UpdatedBy = auditor;
-      }
+      StampEntities(eventData.Context.ChangeTracker, auditor);
 
       return base.SavingChangesAsync(eventData, result, cancellationToken);
     }
 
     #endregion Overrides of SaveChangesInterceptor
+
+    private static void StampEntities<T>(ChangeTracker tracker, T auditor)
+    {
+      var created = tracker.Entries<ICreateAuditable<T>>().Where(x => x.State == EntityState.Added).ToList();
+      foreach (var x in created)
+      {
+        x.Entity.CreatedBy = auditor;
+      }
+
+      var updated = tracker.Entries<IUpdateAuditable<T>>()
+        .Where(x => x.State is EntityState.Added or EntityState.Modified)
+        .ToList();
+      foreach (var x in updated)
+      {
+        x.Entity.UpdatedBy = auditor;
+      }
+    }
   }
 }
